+++ src/App.jsx
@@ -1,9 +1,8 @@
 import Footer from '@components/layouts/Footer';
 import Header from '@components/layouts/Header';
-import { Outlet, useOutletContext } from 'react-router-dom';
+import { Outlet } from 'react-router-dom';
 
 export default function App() {
-  useOutletContext()
   return (
     <div>
       <Header />
